Decompile UpdateStageEntities
