+++ src/st/dre/16F90.c
@@ -3,7 +3,7 @@
 
 #include "../update.h"
 
-INCLUDE_ASM("asm/us/st/dre/nonmatchings/16F90", UpdateStageEntities);
+#include "../update_stage_entities.h"
 
 INCLUDE_ASM("asm/us/st/dre/nonmatchings/16F90", func_801973C4);
