+++ src/st/no3/4184C.c
@@ -3,7 +3,7 @@
 
 #include "../update.h"
 
-INCLUDE_ASM("asm/us/st/no3/nonmatchings/4184C", UpdateStageEntities);
+#include "../update_stage_entities.h"
 
 INCLUDE_ASM("asm/us/st/no3/nonmatchings/4184C", func_801C1C80);
